[feature/fix] Added footer choose for some pages, added page template for all posts, increased size of uploaded file, minor page refactoring, added file deletion, added confirming for deletion, increased length of postText in database

+++ src/main/resources/static/scripts/management/messagesManagementService.js
@@ -39,9 +39,11 @@ async function markMessage(id, status) {
 
 async function deleteMessage(id) {
     const url = `${serverURL}/messages_management/${id}`;
-    const response = await fetch(url, {method: 'DELETE'});
-    if (!await handleError(response)) {
-        alert(`Message was deleted successfully`);
+    if (confirm('Are you sure you want to delete this message?')) {
+        const response = await fetch(url, {method: 'DELETE'});
+        if (!await handleError(response)) {
+            alert(`Message was deleted successfully`);
+        }
     }
 }
