+++ src/main/resources/static/scripts/management/postManagementService.js
@@ -7,17 +7,18 @@ async function retrievePostFromServer(id) {
     if (!await handleError(response)) {
         const post = JSON.parse(await response.text());
         const user = await retrieveUserFromServer(post.writerID);
-        return createPost(post.postID, post.postHeader, post.postText, user.login, user.id);
+        return createPost(post.postID, post.postHeader, post.postText, user.login, post.writerID, post.postPhotoPath);
     }
 }
 
-function createPost(id, postHeader, postText, writerLogin, writerID) {
+function createPost(id, postHeader, postText, writerLogin, writerID, photoPath) {
     return {
         postID: id,
         postHeader: postHeader,
         postText: postText,
         writerLogin: writerLogin,
-        writerID: writerID
+        writerID: writerID,
+        postPhotoPath: photoPath
     };
 }
 
@@ -27,6 +28,7 @@ function setPostDataToForm(post) {
     document.getElementById('post_management_update_text_input').value = post.postText;
     document.getElementById('post_management_update_writer_input').value = post.writerLogin;
     document.getElementById('post_management_update_writerID_input').value = post.writerID;
+    document.getElementById('post_management_update_photo_input').value = post.postPhotoPath;
 }
 
 function retrieveDataFromForm() {
@@ -35,8 +37,9 @@ function retrieveDataFromForm() {
     const text = document.getElementById('post_management_update_text_input').value;
     const writerLogin = document.getElementById('post_management_update_writer_input').value;
     const writerID = document.getElementById('post_management_update_writerID_input').value;
+    const postPhotoPath = document.getElementById('post_management_update_photo_input').value;
 
-    return createPost(id, header, text, writerLogin, writerID);
+    return createPost(id, header, text, writerLogin, writerID, postPhotoPath);
 }
 
 export function manageModal() {
@@ -63,10 +66,8 @@ export function setUpdateEventListeners() {
 
 export function setDeleteEventListener() {
     const buttons = document.getElementsByClassName('delete-button');
-    const deleteButtonListener = e => {
-        deletePost(resolveElementID(e.target.id))
-            .then(() => window.location.reload());
-    }
+    const deleteButtonListener = e => deletePost(resolveElementID(e.target.id))
+        .then(() => window.location.reload());
 
     for (let button of buttons) {
         button.addEventListener('click', deleteButtonListener);
@@ -75,22 +76,29 @@ export function setDeleteEventListener() {
 
 export function setSaveEventListener() {
     const button = document.getElementById('save_button');
-    button.addEventListener('click', updatePost);
+    const saveButtonListener = () => updatePost().then(() => window.location.reload());
+    button.addEventListener('click', saveButtonListener);
 }
 
 async function deletePost(id) {
     const url = `${serverURL}/post_management/${id}`;
-    const response = await fetch(url, {method: 'DELETE'});
-    if (!await handleError(response)) {
-        const deletedPost = JSON.parse(await response.text());
-        alert(`Post with header "${deletedPost.postHeader}" was successfully deleted`);
+    if (confirm('Are you sure you want to delete this post?')) {
+        const response = await fetch(url, {method: 'DELETE'});
+        if (!await handleError(response)) {
+            const deletedPost = JSON.parse(await response.text());
+            alert(`Post with header "${deletedPost.postHeader}" was successfully deleted`);
+        }
     }
 }
 
 async function updatePost() {
     const postToUpdate = retrieveDataFromForm();
     const url = `${serverURL}/post_management/${postToUpdate.postID}`;
-    const response = await fetch(url, {method: 'PATCH', body: JSON.stringify(postToUpdate)});
+    const response = await fetch(url, {
+        method: 'PATCH', body: JSON.stringify(postToUpdate), headers: {
+            'Content-Type': 'application/json'
+        }
+    });
     if (!await handleError(response)) {
         const updatedPost = JSON.parse(await response.text());
         alert(`Post "${updatedPost.postHeader}" was updated successfully`);
