+++ src/main/resources/static/scripts/management/userManagementService.js
@@ -10,35 +10,39 @@ export function manageModal() {
     };
 }
 
+function createUser(id, name, login, userRole, postsNumber) {
+    return {
+        userID: id,
+        login: login,
+        name: name,
+        userRole: userRole,
+        postsNumber: postsNumber
+    };
+}
+
 function retrieveUserDetailsFromForm() {
     const id = document.getElementById("user_management_update_id_input").value;
     const name = document.getElementById("user_management_update_name_input").value;
     const login = document.getElementById("user_management_update_login_input").value;
     const userRole = document.getElementById("user_management_update_role_input").value;
-    return createUser(id, name, login, userRole);
+    const postsNumber = document.getElementById("user_management_update_posts_input").value;
+    return createUser(id, name, login, userRole, postsNumber);
 }
 
-function setUserDetailsToForm(id, name, login, userRole) {
+function setUserDetailsToForm(id, name, login, userRole, postsNumber) {
     document.getElementById("user_management_update_id_input").value = id;
     document.getElementById("user_management_update_name_input").value = name;
     document.getElementById("user_management_update_login_input").value = login;
     document.getElementById("user_management_update_role_input").value = userRole;
-}
-
-function createUser(id, name, login, userRole) {
-    return {
-        userid: id,
-        login: login,
-        name: name,
-        userRole: userRole
-    };
+    document.getElementById("user_management_update_posts_input").value = postsNumber;
 }
 
 function getUserDetailsFromRow(userId) {
     const userName = document.getElementById(`name-${userId}`).innerHTML;
     const userLogin = document.getElementById(`login-${userId}`).innerHTML;
     const userRole = document.getElementById(`userRole-${userId}`).innerHTML;
-    return createUser(userId, userName, userLogin, userRole);
+    const postsNumber = document.getElementById(`postsNumber-${userId}`).innerHTML;
+    return createUser(userId, userName, userLogin, userRole, postsNumber);
 }
 
 export function setUpdateEventListeners() {
@@ -47,7 +51,7 @@ export function setUpdateEventListeners() {
         const buttonId = e.target.id;
         const userId = buttonId.split("-")[1];
         const user = getUserDetailsFromRow(userId);
-        setUserDetailsToForm(user.userid, user.name, user.login, user.userRole);
+        setUserDetailsToForm(user.userID, user.name, user.login, user.userRole, user.postsNumber);
         document.getElementById('user_management_update_form').style.display = 'block'
     }
 
@@ -73,11 +77,11 @@ export function setDeleteEventListeners() {
 
 export async function updateUserData() {
     const user = retrieveUserDetailsFromForm();
-    const url = `${serverURL}/user_management/${user.userid}?name=${user.name}&login=${user.login}&userRole=${user.userRole}`;
+    const url = `${serverURL}/user_management/${user.userID}`;
 
     const response = await fetch(url, {
         method: 'PATCH', body: JSON.stringify(user), headers: {
-            'ContentType': 'application/json'
+            'Content-Type': 'application/json'
         }
     });
     if (!await handleError(response)) {
@@ -89,11 +93,13 @@ export async function updateUserData() {
 }
 
 async function deleteUser(user) {
-    const url = `${serverURL}/user_management/${user.userid}`;
-    const response = await fetch(url, {method: 'DELETE'});
-    if (!await handleError(response)) {
-        const deletedUser = JSON.parse(await response.text())
-        alert(`User ${deletedUser.login} was deleted successfully`);
+    const url = `${serverURL}/user_management/${user.userID}`;
+    if (confirm(`Are you sure you want to delete user '${user.login}'`)) {
+        const response = await fetch(url, {method: 'DELETE'});
+        if (!await handleError(response)) {
+            const deletedUser = JSON.parse(await response.text())
+            alert(`User ${deletedUser.login} was deleted successfully`);
+        }
     }
 }
